% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupGluster.R
\name{collateGroupGluster}
\alias{collateGroupGluster}
\title{Binds groupGluster}
\usage{
collateGroupGluster(...)
}
\arguments{
\item{...}{groupGluster objects.}
}
\value{
Analogous to gluster, groupGluster simpky returns lists in the name of slides, each list being corresponding gluster object.
}
\description{
Extends the c() for groupGluster object
}
\details{
Fits cfGMM models to each marker channel in a matrix of marker channels for multiple slides
}
