% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.R
\name{convCheck}
\alias{convCheck}
\title{Checks convergence of gluster/groupGluster object}
\usage{
convCheck(gluster.fit, out.all = FALSE, return.result = FALSE)
}
\arguments{
\item{gluster.fit}{gluster or groupGluster object.}

\item{out.all}{Whether to print out the full convergence result for all markers and slides, or just summary of non-convergence.}

\item{return.result}{Whether to return the convergence result into an object.}
}
\value{
Returns simple printout or dataframe of convergence result, depending on argument input.
}
\description{
This needs to be run before any plot of fitted result being made.
}
\details{
convergence check to gluster or batch gluster fits
}
