% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.R
\name{diag_contrast}
\alias{diag_contrast}
\title{Diagnostic plot - contrast}
\usage{
diag_contrast(
  fit1,
  fit2,
  marker = 1,
  component = 2,
  title = "Contrastive Diagnostic Plot",
  fit.names = c("fit1", "fit2"),
  boundaries = NULL,
  ...
)
}
\arguments{
\item{fit1}{A groupGluster object.}

\item{fit2}{Another groupGluster object, possibly that is a refit of fit1 with updated constraints.}

\item{marker}{Marker to be plotted, defaults to the first one.}

\item{title}{Title of plot.}

\item{fit.names}{Names of fit objects. Default to c("fit1", "fit2")}

\item{...}{Arguments passed to plot.gluster}

\item{subBatch}{subBatch to be plotted, defaults to the first one.}

\item{add.table}{logical indicating whether to add table of fit2 parameter values.}
}
\description{
Diagnostic plot - contrast
}
\details{
Plot a histogram of a gluster model. Plots one model.
Takes a gluster object and plots the histogram with the fitted model and parameter values.
}
