% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{hist_sr_constrast_inner}
\alias{hist_sr_constrast_inner}
\title{Fits a gamma mixture model to aid clustering of mIF imaging data}
\usage{
hist_sr_constrast_inner(
  fit1,
  fit2,
  marker = 1,
  subBatch = 1,
  title = NULL,
  ...
)
}
\arguments{
\item{fit1}{A gluster object.}

\item{fit2}{Another gluster object, possibly that is a refit of fit1 with updated constraints.}

\item{marker}{Marker to be plotted, defaults to the first one.}

\item{subBatch}{subBatch to be plotted, defaults to the first one.}

\item{title}{Title of plot.}

\item{...}{Arguments passed to plot.gluster}
}
\description{
Fits a gamma mixture model to aid clustering of mIF imaging data
}
\details{
Plot a histogram of a gluster model. Plots one model.
Takes a gluster object and plots the histogram with the fitted model and parameter values.
}
