% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{kappaGroupGluster}
\alias{kappaGroupGluster}
\title{Compare binary classification result using Cohen's Kappa}
\usage{
kappaGroupGluster(..., standard, batch)
}
\arguments{
\item{...}{Matrices to compare to standard labels. Rows are cells, columns are markers.}

\item{standard}{A matrix indicating gold/silver standard positive cells.}

\item{batch}{Split kappa computation by this variable. Similar to slide_id.}
}
\description{
This function can compare the agreement of classification result against a gold/silver standard.
The result can be returned as in either a data frame of Cohen's Kappa or a side-by-side box plot.
}
\details{
Plot a histogram of a gluster model. Plots one model.
Takes a gluster object and plots the histogram with the fitted model and parameter values.
}
