% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R, R/groupGluster.R
\name{plot.groupGluster}
\alias{plot.groupGluster}
\title{Fits a gamma mixture model to aid clustering of mIF imaging data}
\usage{
\method{plot}{groupGluster}(
  x,
  marker = 1,
  component = 2,
  diagnostic = TRUE,
  histogram = FALSE,
  title = NULL,
  color = "grey",
  p = NULL,
  print = TRUE,
  ...
)

\method{plot}{groupGluster}(
  x,
  marker = 1,
  component = 2,
  diagnostic = TRUE,
  histogram = FALSE,
  title = NULL,
  color = "grey",
  p = NULL,
  print = TRUE,
  ...
)
}
\arguments{
\item{x}{A groupGluster object}

\item{marker}{Select which marker to plot. Can be a character or integer.}

\item{component}{Integer specifying which component to plot, 1 is unexpressed nonzero cells, 2 is expressed cells.}

\item{diagnostic}{logical indicating whether to create the diagnostic plot. Default value is TRUE.}

\item{histogram}{logical indicating whether to create the slide histograms.}

\item{title}{Title for the plot. Default is the marker name.}

\item{color}{color for points.}

\item{p}{a ggplot2 object to add to.}

\item{print}{logical whether to display the plot. Default value TRUE.}

\item{...}{Arguments passed to XX}

\item{subBatch}{subBatch to be plotted, defaults to the first one.}

\item{zero_include}{Logical whether the histogram should include zeros. Default is FALSE.}

\item{breaks}{number of breaks for histogram}

\item{boundary}{Vertical line drawn on plot to indicate boundary constraint.}

\item{hist}{Logical indicating whether to plot histogram and base layer of plot.}

\item{diag}{Logical indicating whether to plot interactive diagnostic plot.}

\item{dens}{Logical indicating whether to draw parametric density over histogram.}

\item{tabl}{Logical indicating whether to add table of parameter values.}

\item{col}{Color of density.}
}
\description{
Takes a cfGMM model and forces the right-most component
probabilities to be monotonic increasing with respect
to the input values.
}
\details{
Plot a histogram of a gluster model. Plots one model.
Takes a gluster object and plots the histogram with the fitted model and parameter values.

Various diagnostic and QC plots for groupGluster fits.
}
