% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{get.mode}
\alias{get.mode}
\title{Fits a gamma mixture model to aid clustering of mIF imaging data}
\usage{
get.mode(par.mat)
}
\arguments{
\item{par.mat}{2x2 parameter matrix}
}
\description{
Fits a gamma mixture model to aid clustering of mIF imaging data
}
\details{
Internal Function
Calculate mode from parameter matrix
}
