% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{kappaGroupGluster}
\alias{kappaGroupGluster}
\title{Fits a gamma mixture model to aid clustering of mIF imaging data}
\usage{
kappaGroupGluster(..., standard, batch)
}
\arguments{
\item{...}{Matrices to compare to standard labels.}

\item{standard}{A matrix indicating gold/silver standard positive cells.}

\item{batch}{Split kappa computation by this variable.}
}
\description{
Fits a gamma mixture model to aid clustering of mIF imaging data
}
\details{
Plot a histogram of a gluster model. Plots one model.
Takes a gluster object and plots the histogram with the fitted model and parameter values.
}
