% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupGluster.R
\name{plot.groupGluster}
\alias{plot.groupGluster}
\title{Fits a gamma mixture model to mIF imaging data with multiple slides at one time}
\usage{
\method{plot}{groupGluster}(
  x,
  marker = 1,
  component = 2,
  diagnostic = TRUE,
  histogram = FALSE,
  title = NULL,
  color = "grey",
  p = NULL,
  print = TRUE,
  ...
)
}
\arguments{
\item{x}{A groupGluster object}

\item{marker}{Select which marker to plot. Can be a character or integer.}

\item{component}{Integer specifying which component to plot, 1 is unexpressed nonzero cells, 2 is expressed cells.}

\item{diagnostic}{logical indicating whether to create the diagnostic plot. Default value is TRUE.}

\item{histogram}{logical indicating whether to create the slide histograms.}

\item{title}{Title for the plot. Default is the marker name.}

\item{color}{color for points.}

\item{p}{a ggplot2 object to add to.}

\item{print}{logical whether to display the plot. Default value TRUE.}

\item{...}{Arguments passed to XX}
}
\description{
Takes a cfGMM model and forces the right-most component
probabilities to be monotonic increasing with respect
to the input values.
}
\details{
Various diagnostic and QC plots for groupGluster fits.
}
